Keep popup regression fix within CI file-length guard
The popup routing change itself was fine, but the follow-up diff grew two
already-large Swift source files and one large browser test file enough to
trip the Swift file-length budget guard.

This keeps the behavior unchanged while relocating the popup decision coverage
into a small existing browser test file and compacting the popup-feature helper
wiring so the oversized files stop growing.

Constraint: CI enforces per-file Swift file-length budgets on oversized files
Rejected: Refresh the Swift file-length budget | unnecessary debt acceptance for a narrow follow-up
Confidence: high
Scope-risk: narrow
Reversibility: clean
Directive: Keep future popup routing coverage in small existing browser test files unless the budget is intentionally updated
Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv
Not-tested: Debug app build not run locally
Not-tested: Full CI rerun pending
