+++ cmuxTests/BrowserFindJavaScriptTests.swift
@@ -1,4 +1,6 @@
 import XCTest
+import AppKit
+import WebKit
 
 #if canImport(cmux_DEV)
 @testable import cmux_DEV
@@ -112,5 +114,90 @@ final class BrowserFindJavaScriptTests: XCTestCase {
     func testSearchScriptHandlesLineSeparator() {
         let js = BrowserFindJavaScript.searchScript(query: "test\u{2028}break")
         XCTAssertTrue(js.contains("\\u2028"))
+    }
+}
+
+final class BrowserPopupDecisionTests: XCTestCase {
+    func testLinkActivatedPlainLeftClickDoesNotCreatePopup() {
+        XCTAssertFalse(
+            browserNavigationShouldCreatePopup(
+                navigationType: .linkActivated,
+                modifierFlags: [],
+                buttonNumber: 0
+            )
+        )
+    }
+
+    func testOtherNavigationWithPopupFeaturesCreatesPopup() {
+        XCTAssertTrue(
+            browserNavigationShouldCreatePopup(
+                navigationType: .other,
+                modifierFlags: [],
+                buttonNumber: 0,
+                popupFeaturesWereSpecified: true,
+                currentEventType: .keyDown,
+                currentEventButtonNumber: 0
+            )
+        )
+    }
+
+    func testOtherNavigationWithoutPopupFeaturesDoesNotCreatePopup() {
+        XCTAssertFalse(
+            browserNavigationShouldCreatePopup(
+                navigationType: .other,
+                modifierFlags: [],
+                buttonNumber: 0
+            )
+        )
+    }
+
+    func testOtherNavigationMiddleClickDoesNotCreatePopup() {
+        XCTAssertFalse(
+            browserNavigationShouldCreatePopup(
+                navigationType: .other,
+                modifierFlags: [],
+                buttonNumber: 2
+            )
+        )
+    }
+
+    func testLinkActivatedCmdClickDoesNotCreatePopup() {
+        XCTAssertFalse(
+            browserNavigationShouldCreatePopup(
+                navigationType: .linkActivated,
+                modifierFlags: [.command],
+                buttonNumber: 0
+            )
+        )
+    }
+
+    func testPopupFeaturesAreAbsentWhenAllWindowFeaturesAreNil() {
+        XCTAssertFalse(
+            browserNavigationPopupFeaturesWereSpecified(
+                x: nil,
+                y: nil,
+                width: nil,
+                height: nil,
+                menuBarVisibility: nil,
+                statusBarVisibility: nil,
+                toolbarsVisibility: nil,
+                allowsResizing: nil
+            )
+        )
+    }
+
+    func testPopupFeaturesArePresentWhenWidthIsSpecified() {
+        XCTAssertTrue(
+            browserNavigationPopupFeaturesWereSpecified(
+                x: nil,
+                y: nil,
+                width: NSNumber(value: 640),
+                height: nil,
+                menuBarVisibility: nil,
+                statusBarVisibility: nil,
+                toolbarsVisibility: nil,
+                allowsResizing: nil
+            )
+        )
     }
 }
